requirements.txt
autonet-api
fabric>=2.7.0
pytest